@@ -1,7 +1,7 @@
 [package]
 name = "flathub-stats"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 bitcode = { version = "0.6", features = ["serde"] }
