@@ -1,10 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use std::str::FromStr;
+use std::sync::LazyLock;
 
 use i18n_embed::{
-    fluent::{fluent_language_loader, FluentLanguageLoader},
     DefaultLocalizer, LanguageLoader, Localizer,
+    fluent::{FluentLanguageLoader, fluent_language_loader},
 };
 use icu_collator::{Collator, CollatorOptions, Numeric};
 use icu_provider::DataLocale;
@@ -14,34 +15,30 @@ use rust_embed::RustEmbed;
 #[folder = "i18n/"]
 struct Localizations;
 
-lazy_static::lazy_static! {
-    pub static ref LANGUAGE_LOADER: FluentLanguageLoader = {
-        let loader: FluentLanguageLoader = fluent_language_loader!();
+pub static LANGUAGE_LOADER: LazyLock<FluentLanguageLoader> = LazyLock::new(|| {
+    let loader: FluentLanguageLoader = fluent_language_loader!();
 
-        loader
-            .load_fallback_language(&Localizations)
-            .expect("Error while loading fallback language");
+    loader
+        .load_fallback_language(&Localizations)
+        .expect("Error while loading fallback language");
 
-        loader
-    };
-}
+    loader
+});
 
-lazy_static::lazy_static! {
-    pub static ref LANGUAGE_SORTER: Collator = {
+pub static LANGUAGE_SORTER: LazyLock<Collator> = LazyLock::new(|| {
     let mut options = CollatorOptions::new();
     options.numeric = Some(Numeric::On);
 
     DataLocale::from_str(&LANGUAGE_LOADER.current_language().to_string())
-            .or_else(|_| DataLocale::from_str(&LANGUAGE_LOADER.fallback_language().to_string()))
-            .ok()
-            .and_then(|locale| Collator::try_new(&locale, options).ok())
-            .or_else(|| {
-                let locale = DataLocale::from_str("en-US").expect("en-US is a valid BCP-47 tag");
-                Collator::try_new(&locale, options).ok()
-            })
-            .expect("Creating a collator from the system's current language, the fallback language, or American English should succeed")
-        };
-}
+        .or_else(|_| DataLocale::from_str(&LANGUAGE_LOADER.fallback_language().to_string()))
+        .ok()
+        .and_then(|locale| Collator::try_new(&locale, options).ok())
+        .or_else(|| {
+            let locale = DataLocale::from_str("en-US").expect("en-US is a valid BCP-47 tag");
+            Collator::try_new(&locale, options).ok()
+        })
+        .expect("Creating a collator from the system's current language, the fallback language, or American English should succeed")
+});
 
 #[macro_export]
 macro_rules! fl {
